complete

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use App\Models\Post;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
@@ -92,6 +93,9 @@ class PostController extends Controller
      */
     public function update(UpdatePostRequest $request, Post $post)
     {
+        if (Gate::denies('update',$post)){
+            return abort(401,"You are Account Unauthorized.");
+        }
         $post->title = $request->title;
         $post->slug = Str::slug($request->title);
         $post->description = $request->description;
@@ -117,6 +121,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Gate::denies('update',$post)){
+            return abort(401,"You are Account Unauthorized.");
+        }
         $postTitle = $post->title;
         if (isset($post->featured_image)){
             Storage::delete('public/'.$post->featured_image);
